Rename module `task` to `tasks` to avoid name clash - Also centralize al task related functionalities in the module - Create convinience function to clear exception tracebacks
